fix: fix circular dependency in index, move into utils.ts
affects: @fremtind/jkl-text-input-react
